+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixes infinite loop with `initialValues` in `useForm()`
 - Fixes display of children in `<Field>` when the attribute `options` is undefined and using a custom component
 - Fixes React warnings of duplicate keys with `options` in `<Field>`
+- Returns `initialized` in `useForm()`
 - Updates dependencies
 
 ## v1.0.0